Выделены интерфейсы для хранилищ, добавлен JavaDoc в несколько классов
