+++ README.MD
@@ -1,7 +1,7 @@
 # Программа "ФИТНЕС"
 ### разработчик - Илья Кораблев
 
-Проект разработан с использованием ~Maven~.
+Проект разработан с использованием __Maven__.
 Проект содержит несколько пакетов:
 
 * __models__  - содаржит классы описания сущностей
