+++ src/main/java/ru/list/recover/Main.java
@@ -3,6 +3,10 @@ package ru.list.recover;
 import ru.list.recover.services.Container;
 
 public class Main {
+    
+    /** 
+     * @param args
+     */
     public static void main(String[] args) {
         Container container = new Container();
         container.run();
