+++ src/main/java/ru/list/recover/models/TypeWorkout.java
@@ -14,6 +14,11 @@ public class TypeWorkout {
     private int id;
     private String name;
 
+    
+    /** 
+     * переопределен метод toString
+     * @return String - строковое представление объекта
+     */
     @Override
     public String toString(){
         return String.format("%d %s", this.getId(), this.getName());
