+++ src/main/java/ru/list/recover/models/Workout.java
@@ -14,11 +14,22 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class Workout {
+    /**
+     * уникальный идентификатор трениковки
+     */
     private int id;
+    /**
+     * наименование тренировки
+     */
     private String name;
     private TypeWorkout type;
     private double calories;
 
+    
+    /** 
+     * переопределен метод toString
+     * @return String - строковое представление объекта
+     */
     @Override
     public String toString(){
         return String.format("%d %s (%s)", this.getId(), this.getName(), this.getType().getName());
