+++ src/main/java/ru/list/recover/storages/Repository.java
@@ -2,6 +2,9 @@ package ru.list.recover.storages;
 
 import java.util.List;
 
+/**
+ * 
+ */
 public interface Repository<T> {
 
     boolean insert(T object);
